fix: add back option to publish AMI for 1.31.4
To upgrade our AMIs for customers, we have to be careful that test AMIs
are not auto-released. We do this by using a Kubernetes version
different from the ones our customers have deployed (for now).

We have production customers (e.g. BentoBox) on 1.31.4, so we want to
explicitly build an AMI for & launch a Valet cluster with k8s version
1.31.5 to avoid new Bento nodes getting scheduled onto our new untested
AMI.

X-Related-Ticket-Id: PLA-204 in Linear